ref: rename `CustomComponents.js` to `CustomMapComponents.js`as suggested by reviewer
